+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [ Unreleased ]
 
+### Add
+- Text animation to display entire tool names when they are too long
+
 ### Fix
 - Indicators collection on web pages containing private IP addresses